user mode, printf

+++ stdlib/stdio.h
@@ -2,6 +2,6 @@
 
 int printf(const char *format, ...);
 
-ssize_t getline(char **lineptr, size_t *n);
+int64_t getline(char **lineptr, size_t *n);
 
 void perror(const char *s);
