+++ stdlib/unistd.c
@@ -4,7 +4,7 @@
 #define SYS_read 0
 #define SYS_write 1
 
-extern int syscall(uint64_t nr, ...);
+extern int64_t syscall(uint64_t nr, ...);
 
 int64_t read(int fd, void *buf, size_t count) {
   if (count == 0) return 0;
